requirements.txt
PyPDF2==3.0.1
pdfplumber>=0.11.0
Pillow>=10.3.0
reportlab>=4.2.0
opencv-python>=4.10.0
numpy>=1.26.0
